Major update; features added, bugs fixed
* (C++, .NET) Exponentiation in all contexts now takes unsigned integer
* values (BigUInt or uint64_t)
* (C++) Fix a bug in MemoryPool::get_for_uint64_count() caused by
* pools_.begin() dereferencing to junk when pools_ is empty
* (C++, .NET missing still) Create functions
* exponentiate_poly_polymod_coeffmod, exponentiate_uint_mod,
* poly_eval_poly_polymod_coeffmod, poly_eval_uint_mod
* (C++, .NET) Remove functions exponentiate_poly and exponentiate_uint
* from public API
* (C++) Change some throws from out_of_range to invalid_argument
* (C++, .NET) Change the default prime for n=8192 to be just below a
* multiple of 64 bits (it was just ABOVE a multiple of 64 bits earlier,
* resulting in longer ciphertexts)
* (C++, .NET) select_parameters now takes also default_noise_max_deviation
* as argument
* (C++, .NET) Add a static variable and getter for
* default_noise_max_deviation to ChooserEvaluator
* (C++, .NET) Decryptor::decrypt() now sets the size of the output to have
* coefficients modulo the plaintext modulus
* (C++, .NET) Decryptor::Decryptor() now takes the power as uint64_t
* (C++, .NET) KeyGenerator::generate() now takes the power as uint64_t
* (C++, .NET) Remove unit tests related to exponentiate_poly,
* exponentiate_uint
* (C++, .NET) Add classes PolyCRTBuilder and included polycrt.h to seal.h
* (C++, .NET) Add examples demonstrating PolyCRTBuilder
* (C++, .NET) Make PolyCRTBuilder::Compose() somewhat thread-safe by using
* a local MemoryPool. Need to prepare_all_slots() first.
* (C++) Change code of PolyCRTBuilder::prepare_slot() so that it doesn't
* rescale the polynomials is a flag is set
* (C++) Fix a bug in util::modulo_uint_inplace(...) that caused incorrect
* behavior when modulus had length a multiple of 64 bits (and > 64 bits)
* (C++, .NET) Improve code of PolyCRTBuilder, fix a minor bug, improve
* comments
* (.NET) Fix a bug where GC would dispose of .NET objects too quickly
* without realizing that C++ pointers are still pointing to their wrapped
* objects
* (C++) Change default_parameter_options_ to use better primes
* (C++) default_parameter_options_ now recommends 191 bit coeff modulus
* (earlier was 127) for 1x^4096+1 poly modulus
* (C++) Further changes to recommended parameters in chooser.cpp
* (C++, .NET) Minor changes to examples; improved commentary
* (C++, .NET) Improve functions in utilities.cpp/UtilitiesWrapper.cpp
* (C++) Change makefile to include polycrt.cpp

+++ SEALNETTest/UtilitiesWrapper.cs
@@ -84,10 +84,11 @@ namespace SEALNETTest
             Assert.AreEqual(1, Utilities.EstimateLevelMax(parms));
         }
 
+        /*
         [TestMethod]
         public void ExponentiateUIntNET()
         {
-            int exponent;
+            System.UInt64 exponent;
             var bigUInt = new BigUInt();
             var expUInt = new BigUInt();
 
@@ -165,7 +166,7 @@ namespace SEALNETTest
         [TestMethod]
         public void ExponentiatePolyNET()
         {
-            int exponent;
+            System.UInt64 exponent;
             var bigPoly = new BigPoly();
             var expPoly = new BigPoly();
 
@@ -239,6 +240,7 @@ namespace SEALNETTest
             expPoly = Utilities.ExponentiatePoly(bigPoly, exponent);
             Assert.IsTrue(expPoly.ToString() == "30684B4BF0E5E24DC014B5AC590720EB9AD08D8DF6046110F8F5AF53B8A61F969267EC1Dx^123");
         }
+        */
 
         [TestMethod]
         public void PolyEvalPolyNET()
